Further working on subscriber
Hopefully fixed ws compilation error, but still having error when trying to run subinit.cpp. Error may be that a library is not being included, but I've checked my libraries included relative to tutorials and they should be fine, as well as my cmake which looks fine to me. I may be including the libraries incorrectly in cmake, will be next thing I look at.